requirements.txt
phidata
python-dotenv
groq
wikipedia
asyncio
crawl4ai
baidusearch
pycountry
streamlit
flask